refactor(cfd-1d): Decompose serpentine/mod.rs — extract types leaf module
Part of ATLAS-ARCH-007.
